written java doc in each file and updated the readme.md in docs
