+++ src/main/java/g/gui/Main.java
@@ -1,8 +1,8 @@
-package duke.gui;
+package g.gui;
 
-import duke.Duke;
 import java.io.IOException;
 
+import g.G;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -14,7 +14,7 @@ import javafx.stage.Stage;
  */
 public class Main extends Application {
 
-    private Duke duke = new Duke();
+    private G g = new G();
 
     @Override
     public void start(Stage stage) {
@@ -23,7 +23,7 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);  // inject the Duke instance
+            fxmlLoader.<MainWindow>getController().setDuke(g);  // inject the G instance
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
