+++ src/main/java/g/storage/Storage.java
@@ -1,15 +1,20 @@
-package duke.storage;
+package g.storage;
+
 import java.io.*;
 import java.util.ArrayList;
+import g.tasks.*;
 
-import duke.tasks.*;
-
+/**
+ * Handles the reading and writing of tasks to a storage file.
+ * This class ensures task data is persisted across application sessions.
+ */
 public class Storage {
     private final String filePath;
 
     /**
-     * 
-     * @param filePath
+     * Initializes a Storage object and ensures the storage file exists.
+     *
+     * @param filePath The file path where tasks will be stored.
      */
     public Storage(String filePath) {
         this.filePath = filePath;
@@ -17,21 +22,21 @@ public class Storage {
         File directory = new File(file.getParent());
 
         try {
-            if (!file.exists()) {
-                file.createNewFile();
-            }
-
             if (!directory.exists()) {
-                directory.mkdirs();
+                directory.mkdirs(); // Create directory if it does not exist
+            }
+            if (!file.exists()) {
+                file.createNewFile(); // Create file if it does not exist
             }
         } catch (IOException e) {
-            System.out.println("Error intializing storage file.");
+            System.out.println("Error initializing storage file.");
         }
     }
 
     /**
-     * 
-     * @return
+     * Loads tasks from the storage file into an ArrayList.
+     *
+     * @return An ArrayList of Task objects loaded from the file.
      */
     public ArrayList<Task> load() {
         ArrayList<Task> tasks = new ArrayList<>();
@@ -41,32 +46,31 @@ public class Storage {
             while ((line = br.readLine()) != null) {
                 String[] parts = line.split(" \\| ");
                 switch (parts[0]) {
-                    case "T" :
+                    case "T":
                         tasks.add(new Todo(parts[2], parts[1].equals("1")));
                         break;
                     case "D":
                         tasks.add(new Deadline(parts[2], parts[3], parts[1].equals("1")));
                         break;
                     case "E":
-                        tasks.add(new Event(parts[2], parts[3], parts[4],  parts[1].equals("1")));
+                        tasks.add(new Event(parts[2], parts[3], parts[4], parts[1].equals("1")));
                         break;
                 }
             }
-            br.close();
-
         } catch (IOException e) {
             System.out.println("Error reading from file: " + e.getMessage());
-        } 
+        }
         return tasks;
     }
 
     /**
-     * 
-     * @param tasks
+     * Saves the current task list to the storage file.
+     *
+     * @param tasks The list of tasks to be saved.
      */
     public void save(ArrayList<Task> tasks) {
-        try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.filePath))){
-            for (Task task: tasks) {
+        try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.filePath))) {
+            for (Task task : tasks) {
                 bw.write(task.toFileString());
                 bw.newLine();
             }
